+++ R/03-tryCatchWithErrorsAndWarnings.R
@@ -3,14 +3,9 @@
 #' Catches errors and warnings and forwards them as an alert. If an error occurs, NULL is returned.
 #' If a warning occurs, the result is returned.
 #'
-#' @param expr expression to be evaluated.
-#' @param errorTitle (character) error message title.
-#' @param warningTitle (character) warning message title.
-#' @param alertStyle (character) Either "shinyjs", or "shinyalert". Specifies how an error or a
-#'  warning is given out. If "shinyjs" than shinyjs::alert is used; if "shinyalert" than
-#'  shinyalert::shinyalert is used.
 #' @param silent (logical) if TRUE prevents an alert, instead a warning will be displayed. Use this
 #'  when not applied within a Shiny session.
+#' @inheritParams shinyTools::shinyTryCatch
 #'
 #' @export
 tryCatchWithWarningsAndErrors <- function(expr,
@@ -18,56 +13,15 @@ tryCatchWithWarningsAndErrors <- function(expr,
                                           warningTitle = "",
                                           alertStyle = "shinyjs",
                                           silent = FALSE) {
-  tryCatchMessage <- NULL
+  deprecate_warn("24.05.1",
+                 "DataTools::tryCatchWithWarningsAndErrors()",
+                 "shinyTools::shinyTryCatch()")
 
-  w.handler <- function(w) {
-    # warning handler
-    tryCatchMessage[["text"]] <<- w
-    tryCatchMessage[["text"]] <<-
-      paste0(tryCatchMessage[["text"]], collapse = "\n")
-    tryCatchMessage[["title"]] <<- warningTitle
-    tryCatchMessage[["type"]] <<- "warning"
-    invokeRestart("muffleWarning")
-  }
-
-  e.handler <- function(e) {
-    # error handler
-    tryCatchMessage[["text"]] <<- e$message
-    tryCatchMessage[["title"]] <<- errorTitle
-    tryCatchMessage[["type"]] <<- "error"
-    return(NULL)
-  }
-
-  res <- withCallingHandlers(tryCatch({
-    expr
-  },
-  error = e.handler),
-  warning = w.handler)
-
-  if (!is.null(tryCatchMessage) && !silent) {
-    # give out error or warning
-    switch (
-      alertStyle,
-      "shinyjs" = shinyjs::alert(paste(
-        tryCatchMessage[["title"]],
-        tryCatchMessage[["text"]] %>% as.character(),
-        sep = "\n "
-      )),
-      "shinyalert" = shinyalert::shinyalert(
-        title = tryCatchMessage[["title"]],
-        text = tryCatchMessage[["text"]] %>% as.character(),
-        type = tryCatchMessage[["type"]]
-      )
-    )
-  }
-
-  if (!is.null(tryCatchMessage) && silent) {
-    # give out error or warning
-    warning(paste0(tryCatchMessage[["title"]], tryCatchMessage[["text"]]), call. = FALSE)
-  }
-
-  # output result of expr
-  res
+  shinyTryCatch(expr = expr,
+                errorTitle = errorTitle,
+                warningTitle = warningTitle,
+                alertStyle = alertStyle,
+                inShiny = !silent)
 }
 
 # TEST MODULE -------------------------------------------------------------
@@ -121,10 +75,11 @@ tryCatchWithWarningsAndErrors <- function(expr,
 #
 #   observe({
 #     tmpRes <- {
+#       warning("test warning")
 #       stop("test error")
 #       5 + 4
 #     } %>%
-#       tryCatchWithWarningsAndErrors(errorTitle = "Modeling failed", alertStyle = "shinyalert")
+#       tryCatchWithWarningsAndErrors(errorTitle = "Modeling failed", alertStyle = "shinyalert", silent = TRUE)
 #     testRes(tmpRes)
 #   }) %>%
 #     bindEvent(input$buttonShinyalertErr)
@@ -143,7 +98,10 @@ tryCatchWithWarningsAndErrors <- function(expr,
 #
 #     p <- p + ggplot2::xlim(c(3, 8))
 #
-#     print(p) %>%
+#     {
+#       warning("Test warning")
+#       print(p)
+#       } %>%
 #       tryCatchWithWarningsAndErrors(errorTitle = "Plotting failed: ",
 #                                     warningTitle = "Warning in plotting: ",
 #                                     alertStyle = "shinyalert")